requirements.txt
Django==1.7.3
Pillow==2.7.0
argparse==1.2.1
django-nose==1.3
factory-boy==2.4.1
libsass==0.6.2
mysqlclient==1.3.4
nose==1.3.4
ply==3.4
python-dateutil==2.4.1
pytz==2014.10
requests==2.5.1
six==1.9.0
slimit==0.8.1
wsgiref==0.1.2
